2022-12-27

+++ finalproject/src/main/java/com/kh/sixmusic/product/model/vo/Product.java
@@ -18,7 +18,7 @@ public class Product {
 	private int price;//	PRODUCT_PRICE	상품_가격
 	private int quantity;//	PRODUCT_QUANTITY	상품_수량
 	private String status;//	STATUS	상태값(Y/N)
-	private int rating;//	REVIEW 테이블의  RATING의 평균 값
+	private Double rating;//	REVIEW 테이블의  RATING의 평균 값
 	private int count;//	REVIEW 의 갯수 값
 	private String filePath;
 	private String changeName;
